+++ uforwarder/src/test/java/com/uber/data/kafka/consumerproxy/worker/limiter/LongFixedInflightLimiterTest.java
@@ -1,12 +1,8 @@
 package com.uber.data.kafka.consumerproxy.worker.limiter;
 
 import com.uber.fievel.testing.base.FievelTestBase;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.awaitility.Awaitility;
@@ -192,120 +188,5 @@ public class LongFixedInflightLimiterTest extends FievelTestBase {
     permit1.get().complete(InflightLimiter.Result.Succeed);
     permit3 = inflightLimiter.tryAcquire();
     Assert.assertTrue(permit3.isPresent());
-  }
-
-  @Test
-  public void testAcquireAsync() throws ExecutionException, InterruptedException {
-    CompletableFuture<InflightLimiter.Permit> futurePermit1 = inflightLimiter.acquireAsync();
-    Assert.assertTrue(futurePermit1.isDone());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-    CompletableFuture<InflightLimiter.Permit> futurePermit2 = inflightLimiter.acquireAsync();
-    Assert.assertTrue(futurePermit2.isDone());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-    CompletableFuture<InflightLimiter.Permit> futurePermit3 = inflightLimiter.acquireAsync();
-    Assert.assertFalse(futurePermit3.isDone());
-    Assert.assertEquals(1, inflightLimiter.getMetrics().getAsyncQueueSize());
-    Assert.assertTrue(futurePermit2.get().complete(InflightLimiter.Result.Succeed));
-    Assert.assertTrue(futurePermit3.isDone());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-  }
-
-  @Test
-  public void testCloseRelease() {
-    List<CompletableFuture<InflightLimiter.Permit>> permits = new ArrayList<>();
-    for (int i = 0; i < 3; ++i) {
-      permits.add(inflightLimiter.acquireAsync());
-    }
-    int done = 0;
-    for (CompletableFuture<InflightLimiter.Permit> permit : permits) {
-      done += permit.isDone() ? 1 : 0;
-    }
-    Assert.assertEquals(2, done);
-    inflightLimiter.close();
-    done = 0;
-    for (CompletableFuture<InflightLimiter.Permit> permit : permits) {
-      done += permit.isDone() ? 1 : 0;
-    }
-    Assert.assertEquals(3, done);
-  }
-
-  @Test
-  public void testAcquireSucceedAndCancel() {
-    CompletableFuture<InflightLimiter.Permit> futurePermit = inflightLimiter.acquireAsync();
-    Assert.assertFalse(futurePermit.cancel(false));
-  }
-
-  @Test
-  public void testAcquireFailedAndCancel() throws ExecutionException, InterruptedException {
-    for (int i = 0; i < 2; ++i) {
-      inflightLimiter.acquireAsync();
-    }
-    CompletableFuture<InflightLimiter.Permit> futurePermit = inflightLimiter.acquireAsync();
-    Assert.assertEquals(1, inflightLimiter.getMetrics().getAsyncQueueSize());
-    futurePermit.cancel(false);
-    Assert.assertTrue(futurePermit.isDone());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-  }
-
-  @Test
-  public void testCompletePermitDequeFuturePermit()
-      throws ExecutionException, InterruptedException {
-    CompletableFuture<InflightLimiter.Permit> acquiredPermit = null;
-    for (int i = 0; i < 2; ++i) {
-      acquiredPermit = inflightLimiter.acquireAsync();
-    }
-    CompletableFuture<InflightLimiter.Permit> futurePermit = inflightLimiter.acquireAsync();
-    Assert.assertEquals(1, inflightLimiter.getMetrics().getAsyncQueueSize());
-    acquiredPermit.get().complete(InflightLimiter.Result.Succeed);
-    Assert.assertTrue(futurePermit.isDone());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-  }
-
-  @Test
-  public void testCancelCompletedFuture() throws ExecutionException, InterruptedException {
-    CompletableFuture<InflightLimiter.Permit> acquiredPermit = new CompletableFuture();
-    acquiredPermit.cancel(false);
-    Assert.assertTrue(acquiredPermit.isCancelled());
-    Assert.assertFalse(acquiredPermit.complete(InflightLimiter.NoopPermit.INSTANCE));
-    Assert.assertTrue(acquiredPermit.isDone());
-    Assert.assertTrue(acquiredPermit.isCancelled());
-    Assert.assertTrue(acquiredPermit.isCompletedExceptionally());
-  }
-
-  @Test
-  public void testCompletePermitCompletableFuture() {
-    for (int i = 0; i < 2; ++i) {
-      inflightLimiter.acquireAsync();
-    }
-    CompletableFuture<InflightLimiter.Permit> futurePermit = inflightLimiter.acquireAsync();
-    Assert.assertFalse(futurePermit.isDone());
-    futurePermit.completeExceptionally(new CancellationException());
-    AtomicBoolean complected = new AtomicBoolean(false);
-    futurePermit.complete(
-        new InflightLimiter.AbstractPermit() {
-          @Override
-          protected boolean doComplete(InflightLimiter.Result result) {
-            return complected.compareAndSet(false, true);
-          }
-        });
-    Assert.assertEquals(true, complected.get());
-  }
-
-  @Test
-  public void testCompletePermitDequeCompletedFuturePermit()
-      throws ExecutionException, InterruptedException {
-    CompletableFuture<InflightLimiter.Permit> acquiredPermit = null;
-    for (int i = 0; i < 2; ++i) {
-      acquiredPermit = inflightLimiter.acquireAsync();
-    }
-    CompletableFuture<InflightLimiter.Permit> futurePermit = inflightLimiter.acquireAsync();
-    Assert.assertFalse(futurePermit.isDone());
-    futurePermit.completeExceptionally(new CancellationException()); // complete a future permit
-    Assert.assertTrue(futurePermit.isDone());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-    Assert.assertEquals(0, inflightLimiter.getMetrics().availablePermits());
-    acquiredPermit.get().complete(InflightLimiter.Result.Succeed); // it should return permit
-    Assert.assertEquals(0, inflightLimiter.getMetrics().getAsyncQueueSize());
-    Assert.assertEquals(1, inflightLimiter.getMetrics().availablePermits());
   }
 }
